split tests into separate file
